% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal2norm}
\alias{cal2norm}
\title{Calculate group norms}
\usage{
cal2norm(u, g_idx, type)
}
\arguments{
\item{u}{\code{C2*gamma.coef}, \code{gamma.coef} is the estimated node coefficient vector,
\code{C2} matrix is the output from
function \code{get_tree_object()}, which gives the weights of the groups.}

\item{g_idx}{Group structure matrix defined by the \code{C2} matrix.
See details in \code{get_tree_object()}.}

\item{type}{If \code{type == 1}, return sum of group norms; else return individual
norm for each group.}
}
\value{
Sum of group norms or individual group norms.
}
\description{
Function to output group norms on the gamma coefficients
based on \code{g_idx} and \code{C2} matrix.
}
