% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{getaggr}
\alias{getaggr}
\title{Generate aggregated features}
\usage{
getaggr(TSLA.object, X_2, X_2.org, lambda.index, alpha.index)
}
\arguments{
\item{TSLA.object}{A fit output from \code{TSLA.fit()},
or the \code{TSLA.fit} object in \code{cv.TSLA()}.}

\item{X_2}{Expanded design matrix in matrix form.}

\item{X_2.org}{Original design matrix in matrix form.}

\item{lambda.index}{Index of the \eqn{\lambda} value selected.}

\item{alpha.index}{Index of the \eqn{\alpha} value selected. The \eqn{\alpha} is
the tuning parameter for generalized lasso penalty.}
}
\value{
A data.frame of the aggregated feature.
\item{dataset}{aggregated features.}
}
\description{
Function that generates aggregated features based on the TSLA output.
}
