% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict_cvTSLA}
\alias{predict_cvTSLA}
\title{Prediction from cross validation}
\usage{
predict_cvTSLA(
  object,
  X_1_new = NULL,
  X_2_new,
  type = c("response", "link"),
  ...
)
}
\arguments{
\item{object}{A fit output from \code{cv.TSLA().}}

\item{X_1_new}{New unpenalized features in matrix form.}

\item{X_2_new}{New binary features in matrix form.}

\item{type}{Two options: "response" or "link". The two options only
differ for \code{family="logit"}.}

\item{...}{Other parameters.}
}
\value{
Predictions.
}
\description{
A convenient function to get prediction from the
selected tuning parameters by cross validation.
}
