% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict_TSLA}
\alias{predict_TSLA}
\title{Prediction from TSLA with new data}
\usage{
predict_TSLA(
  object,
  X_1_new = NULL,
  X_2_new,
  type = c("response", "link"),
  ...
)
}
\arguments{
\item{object}{A fit output from \code{TSLA.fit()}.}

\item{X_1_new}{New unpenalized features in matrix form.}

\item{X_2_new}{New binary features in matrix form.}

\item{type}{Two options: "response" or "link". The two options only
differ for \code{family="logit"}.}

\item{...}{Other parameters.}
}
\value{
Predictions. The first dimension is indexed by the observation,
the second dimension is indexed by \eqn{\lambda},
and the third dimension is indexed by \eqn{\alpha}.
}
\description{
Generate prediction for the response.
}
