% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegressionExample}
\alias{RegressionExample}
\title{Synthesic for the regression example}
\format{
List containing the following elements:
\describe{
  \item{tree.org}{Original tree structure with 42 leaf nodes and 5 different levels.}
  \item{x.org}{Original design matrix with 42 binary features and 400 observations.}
  \item{y}{Continuous response of length 400.}
}
}
\usage{
data(RegressionExample)
}
\description{
Synthetic data used to illustrate how to use TSLA with regression.
}
\keyword{data}
