% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSLA-package.R
\docType{package}
\name{TSLA-package}
\alias{TSLA-package}
\title{Tree-Guided Rare Feature Selection and Logic Aggregation}
\description{
This package provides functions and visualization tools for fitting the
Tree-Guided Rare Feature Selection and Logic Aggregation model.
}
\author{
Jianmin Chen \email{jianminc000@gmail.com}, Kun Chen
}
